WIP for SSO from govmomi
Working on using the LoginByToken call to Govmomi instead of the Login call.
This requires using a separate client to keep track of session information.
Building off of the WIP done by dougm mentioned in #766.  The new client in sts
is issued a token and works.  In conjunction with another WIP in vic.